Bring up to date with further placement changes
