+++ osu.Game.Rulesets.Osu/Edit/SliderCompositionTool.cs
@@ -3,7 +3,7 @@
 
 using osu.Game.Rulesets.Edit;
 using osu.Game.Rulesets.Edit.Tools;
-using osu.Game.Rulesets.Osu.Edit.Masks;
+using osu.Game.Rulesets.Osu.Edit.Masks.Slider;
 using osu.Game.Rulesets.Osu.Objects;
 
 namespace osu.Game.Rulesets.Osu.Edit
